Improve TagDiff, giving it 3-array and 3-prefix options.
The original two arrays and prefixes exist unchanged, + to add
tags (implied if not present) and - to remove them. There is now
also =, to reset the tag, removing it from the tag diff entirely.

This is meaningfully distinct from the other two and because tag diffs
are meant to represent staged changes to a TagSet, it's valid to want
a "return to original state" mode that doesn't require you to know what
that original state is.