Remove manual json marshalling and unmarshalling from TagDiff.
I thought it would be nice to have but it turns out that it just makes
things complicated when you try to (ab)use TagDiff for things like sources
which use a different delimiter, and while the structure is perfectly capable
of doing that as long as you don't use delimiter-aware methods, the
serialization using APIString was delimiter-aware so it was munging the
object state. Removing that means it stores a more complex representation
but no longer does that.