fix limit in tag fetching to actually match api's limit
if the limit is higher, that will fool the rest of the function
into thinking the api is returning fewer results because there are
no more results to return, and it will prematurely stop.

A more complete fix would be to use multiple threshholds:
an initial guess at the limit, followed by a whatever number
of queries the first query returned, and a known minimum
limit.