Add /resyncdeleted command.
It's a one-time need so the command remains disabled in this commit.

As part of the api update, you can no longer edit posts which you can't see.
previously you could. so, i need to track the deleted state of posts, which i wasn't
before. this command syncs that state for all posts.

With it, I can now generate two sets of accurate tag counts: one for all posts,
and one for only visible ones, and using that, I can better utilize commands like findtagtypos.