Add settings for max number of artists, characters, and sources to display in a post.
There is currently a hard limit of 10 for each, which is also the default if
they are not specified in the configuration file. You SHOULD pick lower values
for them, because the defaults are wildly overpermissive and will result in
inline query responses that are so large telegram will reject them.