Further sanitize rating sanity checking.
Default behavior now returns an error if an invalid rating is passed,
in addition to just a blank value.

Also, added a new sanitizer function which also accepts "original",
designed to be used for editing posts where "keep the original rating"
is a valid choice. This returns blank, with no error.