Use a dummy target for building.
I hate make.
I hate it so fucking much.

TL;DR fsb was a dummy target, so it was always getting rebuilt
if something depended on it, including during e.g. make install,
which is usually run as root, and it was therefore scattering
root-owned files through the golang build cache and making a huge mess.

Fuck make.